Delete students when deleting a course

+++ LexiconLMS/Controllers/CoursesController.cs
@@ -168,10 +168,41 @@ namespace LexiconLMS.Controllers
         [Authorize(Roles = "Teacher")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Course course = db.Courses.Find(id);
-            db.Courses.Remove(course);
-            db.SaveChanges();
+            bool sucess = DeleteStudents(id);
+
+            if (sucess)
+            {
+                Course course = db.Courses.Find(id);
+                db.Courses.Remove(course);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
+        }
+        private bool DeleteStudents(int id)
+        {
+            ApplicationDbContext newDbContext = new ApplicationDbContext();
+            LexiconLMS.Models.ApplicationUser student;
+            
+            //Find students by courseId    
+            var allStudents = newDbContext.Users.Where(u => u.CourseId == id);
+
+                try
+                {
+                    //Delete students by course
+                    foreach (var students in allStudents.ToList())
+                    {
+                        student = db.Users.Find(students.Id);
+                        db.Users.Remove(student);
+                        //db.SaveChanges();
+                    }
+                    return true;
+                }
+                catch (Exception e)
+                {
+                    return false;
+                }
+            //}
+            
         }
 
         protected override void Dispose(bool disposing)
